Move Rule to separate submodule
